requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pymongo==4.6.1
pandas==2.1.3
python-dotenv==1.0.0
matplotlib==3.8.2
folium==0.14.0
PyQt5==5.15.10
scipy==1.11.4
numpy==1.26.2
seaborn==0.13.0